requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
whisper